Added: Class and passing userid to the methods and returns the invoice.

+++ src/main/java/com/bridgelabz/cabinvoice/InvoiceSummary.java
@@ -4,7 +4,7 @@ public class InvoiceSummary
 {
     double avgFare;
     double totalFare;
-    double numOfRides;
+    int numOfRides;
 
     public InvoiceSummary(int numOfRides, double totalFare) {
         this.numOfRides = numOfRides;
